Refactor _stringify_path to filename_parser.py
